+++ pjsip-apps/src/samples/confsample.c
@@ -1,4 +1,4 @@
-/* $Id: confsample.c 3553 2011-05-05 06:14:19Z nanang $ */
+/* $Id: confsample.c 4537 2013-06-19 06:47:43Z riza $ */
 /* 
  * Copyright (C) 2008-2011 Teluu Inc. (http://www.teluu.com)
  * Copyright (C) 2003-2008 Benny Prijono <benny@prijono.org>
@@ -105,7 +105,7 @@ static pj_bool_t input(const char *title, char *buf, pj_size_t len)
     char *p;
 
     printf("%s (empty to cancel): ", title); fflush(stdout);
-    if (fgets(buf, len, stdin) == NULL)
+    if (fgets(buf, (int)len, stdin) == NULL)
 	return PJ_FALSE;
 
     /* Remove trailing newlines. */
